+++ src/dexterity/localroles/subscriber.py
@@ -1,5 +1,7 @@
 # encoding: utf-8
 
+from OFS.interfaces import IObjectWillBeAddedEvent, IObjectWillBeRemovedEvent
+from zope.lifecycleevent.interfaces import IObjectAddedEvent, IObjectRemovedEvent
 from plone import api
 
 from . import logger
@@ -76,5 +78,31 @@ def related_change_on_addition(obj, event):
     """ Set local roles on related objects after addition """
     fti_config = fti_configuration(obj)
     if 'static_config' not in fti_config:
+        return
+    if event.oldParent and event.newParent and event.oldParent == event.newParent:  # rename
+        return
+    related_role_addition(obj, get_state(obj), fti_config)
+
+
+def related_change_on_moving(obj, event):
+    """ Set local roles on related objects after removal """
+    if IObjectWillBeAddedEvent.providedBy(event) or IObjectWillBeRemovedEvent.providedBy(event):  # not move
+        return
+    fti_config = fti_configuration(obj)
+    if 'static_config' not in fti_config:
+        return
+    if event.oldParent and event.newParent and event.oldParent == event.newParent:  # rename
+        return
+    related_role_removal(obj, get_state(obj), fti_config)
+
+
+def related_change_on_moved(obj, event):
+    """ Set local roles on related objects after addition """
+    if IObjectAddedEvent.providedBy(event) or IObjectRemovedEvent.providedBy(event):  # not move
+        return
+    fti_config = fti_configuration(obj)
+    if 'static_config' not in fti_config:
+        return
+    if event.oldParent and event.newParent and event.oldParent == event.newParent:  # rename
         return
     related_role_addition(obj, get_state(obj), fti_config)