+++ src/dexterity/localroles/tests/test_subscriber.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+import transaction
 import unittest2 as unittest
 from zope import event
 from zope.lifecycleevent import ObjectModifiedEvent
@@ -79,3 +80,22 @@ class TestSubscriber(unittest.TestCase):
         item = self.portal['test']
         # The parent is set
         self.assertDictEqual(get_related_roles(self.portal, item.UID()), {'raptor': set(['Editor'])})
+
+    def test_related_change_on_move(self):
+        add_fti_configuration('testingtype', localroles_config)
+        self.portal.invokeFactory('testingtype', 'test', title="Title")
+        item = self.portal['test']
+        # We need to commit here so that _p_jar isn't None and move will work
+        transaction.savepoint(optimistic=True)
+        # The parent is set by addition subscriber
+        self.assertDictEqual(get_related_roles(self.portal, item.UID()), {'raptor': set(['Editor'])})
+        # We create a folder
+        self.portal.invokeFactory('Folder', 'folder')
+        folder = self.portal['folder']
+        self.assertDictEqual(get_related_roles(folder, item.UID()), {})
+        # We move the item
+        api.content.move(source=item, target=folder)
+        # The old parent is changed
+        self.assertDictEqual(get_related_roles(self.portal, item.UID()), {})
+        # The new parent is changed
+        self.assertDictEqual(get_related_roles(folder, item.UID()), {'raptor': set(['Editor'])})